update: getting projects and news, ogp
