+++ src/styles/app/contact/contact.module.scss
@@ -69,6 +69,7 @@
       height: 2.5rem;
       left: 0;
       margin-top: 2rem;
+      margin-bottom: 1rem;
       // background-color: rgb(233, 246, 250);
       z-index: 1150;
       display: flex;
