+++ src/styles/components/Header.module.scss
@@ -1,6 +1,9 @@
-@import "../global/responsive.scss";
+// ヘッダー用フォント
 @import url("https://fonts.googleapis.com/css2?family=Klee+One:wght@400;600&display=swap");
 
+@import "../global/responsive.scss";
+
+// ヘッダー記述部
 .fixed_header {
   width: 100%;
   position: fixed;
